Floor Recognition/requirements.txt
absl-py==1.0.0
alabaster==0.7.12
anyio==2.2.0
appdirs==1.4.4
argh==0.26.2
argon2-cffi==20.1.0
arrow==0.13.1
asn1crypto==1.4.0
boto==2.49.0
filelock==3.3.1
flake8==3.9.2
Flask==1.1.2
gast==0.5.3
gevent==21.8.0
glob2==0.7
grpcio==1.44.0
html5lib==1.1
idna==3.2
imageio==2.9.0
imagesize==1.2.0
importlib-metadata==4.8.1
Jinja2==2.11.3
jinja2-time==0.2.0
joblib==1.1.0
json5==0.9.6
keras==2.8.0
Keras-Preprocessing==1.1.2
lazy-object-proxy==1.6.0
lxml==4.6.3
mypy-extensions==0.4.3
nest-asyncio==1.5.1
numexpr==2.7.3
numpy==1.19.5
numpydoc==1.1.0
oauthlib==3.2.0
olefile==0.46
openpyxl==3.0.9
pandas==1.1.5
pandocfilters==1.4.3
path==16.0.0
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pip==21.2.4
pkginfo==1.7.1
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.26.0
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.17.2
scikit-learn
scipy
SQLAlchemy
tensorboard
tensorboard-data-server
tensorboard-plugin-wit
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.25.0
testpath==0.5.0
text-unidecode==1.3
textdistance==4.2.1
tinycss==0.4
toml==0.10.2
ujson==4.0.2
unicodecsv==0.14.1
Unidecode==1.2.0
urllib3==1.26.7
wcwidth==0.2.5
wheel==0.37.0
wrapt==1.12.1
xgboost==1.6.0
xlrd==2.0.1
XlsxWriter==3.0.1
